pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentegrity"
version = "0.3.0"
description = "The open standard for AI agent integrity. Evaluate, enforce, and prove that autonomous agents are adversarially coherent, environmentally portable, and verifiably assured."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Cogensec Research" },
]
keywords = [
    "ai-security",
    "ai-agents",
    "agent-integrity",
    "agentegrity",
    "autonomous-agents",
    "adversarial-ai",
    "runtime-security",
    "ai-governance",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

[project.optional-dependencies]
crypto = ["cryptography>=41.0"]
claude = ["claude-agent-sdk>=0.1.0"]
langchain = ["langchain-core>=0.3.0"]
openai-agents = ["openai-agents>=0.1.0"]
crewai = ["crewai>=0.70.0"]
google-adk = ["google-adk>=0.3.0"]
llm = ["anthropic>=0.40.0"]
all = [
    "cryptography>=41.0",
    "claude-agent-sdk>=0.1.0",
    "langchain-core>=0.3.0",
    "openai-agents>=0.1.0",
    "crewai>=0.70.0",
    "google-adk>=0.3.0",
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/requie/agentegrity-framework"
Documentation = "https://github.com/requie/agentegrity-framework/tree/main/spec"
Repository = "https://github.com/requie/agentegrity-framework"
Issues = "https://github.com/requie/agentegrity-framework/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agentegrity"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src/agentegrity",
    "tests",
    "examples",
    "spec",
    "README.md",
    "LICENSE",
    "MANIFESTO.md",
    "agentegrity-glossary.md",
    "pyproject.toml",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
